+++ Sources/RSocketCore/Extensions/Extensions.swift
@@ -18,352 +18,7 @@ import Foundation
 import NIO
 import NIOFoundationCompat
 
-struct ConnectionMIMEType {
-    var metadata: MIMEType
-    var data: MIMEType
-}
-
 // MARK: - Payload Decoder
-
-protocol DecoderProtocol {
-    associatedtype Metadata
-    associatedtype Data
-    mutating func decodedPayload(
-        _ payload: Payload,
-        mimeType: ConnectionMIMEType
-    ) throws -> (Metadata, Data)
-}
-
-protocol SingleMetadataDecodable {
-    func decodeMetadata<Decoder>(
-        using metadataDecoder: Decoder
-    ) throws -> Decoder.Metadata where Decoder: MetadataDecoder 
-}
-
-protocol SingleMetadataEncodable {
-    associatedtype CombinableMetadata = Void
-    static func encodeMetadata<Encoder>(
-        _ metadata: Encoder.Metadata,
-        using metadataEncoder: Encoder
-    ) throws -> Self where Encoder: MetadataEncoder
-    static func encodeMetadata<Encoder>(
-        _ metadata: Encoder.Metadata,
-        using metadataEncoder: Encoder,
-        combinedWith other: CombinableMetadata
-    ) throws -> Self where Encoder: MetadataEncoder
-}
-
-extension SingleMetadataEncodable where CombinableMetadata == Void {
-    static func encodeMetadata<Encoder>(
-        _ metadata: Encoder.Metadata, 
-        using metadataEncoder: Encoder, 
-        combinedWith other: Void
-    ) throws -> Self where Encoder : MetadataEncoder {
-        try encodeMetadata(metadata, using: metadataEncoder)
-    }
-}
-
-extension Data: SingleMetadataDecodable {
-    func decodeMetadata<Decoder>(
-        using metadataDecoder: Decoder
-    ) throws -> Decoder.Metadata where Decoder : MetadataDecoder {
-        try metadataDecoder.decode(from: self)
-    }
-}
-
-extension Optional: SingleMetadataDecodable where Wrapped: SingleMetadataDecodable {
-    func decodeMetadata<Decoder>(
-        using metadataDecoder: Decoder
-    ) throws -> Decoder.Metadata where Decoder : MetadataDecoder {
-        guard let self = self else {
-            throw Error.invalid(message: "Metadata is nil an therefore can not be decoded as \(Decoder.Metadata.self) using \(Decoder.self)")
-        }
-        return try self.decodeMetadata(using: metadataDecoder)
-    }
-}
-
-extension Data: SingleMetadataEncodable {
-    static func encodeMetadata<Encoder>(
-        _ metadata: Encoder.Metadata, 
-        using metadataEncoder: Encoder
-    ) throws -> Data where Encoder : MetadataEncoder {
-        try metadataEncoder.encode(metadata)
-    }
-}
-
-extension Optional: SingleMetadataEncodable where Wrapped: SingleMetadataEncodable {
-    static func encodeMetadata<Encoder>(
-        _ metadata: Encoder.Metadata, 
-        using metadataEncoder: Encoder
-    ) throws -> Self where Encoder : MetadataEncoder {
-        try Wrapped.encodeMetadata(metadata, using: metadataEncoder)
-    }
-}
-
-extension Array: SingleMetadataDecodable where Element == CompositeMetadata {
-    func decodeMetadata<Decoder>(
-        using metadataDecoder: Decoder
-    ) throws -> Decoder.Metadata where Decoder : MetadataDecoder {
-        try self.decodeFirst(using: metadataDecoder)
-    }
-}
-
-extension Array: SingleMetadataEncodable where Element == CompositeMetadata {
-    static func encodeMetadata<Encoder>(
-        _ metadata: Encoder.Metadata, 
-        using metadataEncoder: Encoder
-    ) throws -> Self where Encoder : MetadataEncoder {
-        [try CompositeMetadata.encoded(metadata, using: metadataEncoder)]
-    }
-    static func encodeMetadata<Encoder>(
-        _ metadata: Encoder.Metadata, 
-        using metadataEncoder: Encoder, 
-        combinedWith other: Self
-    ) throws -> [CompositeMetadata] where Encoder : MetadataEncoder {
-        other + CollectionOfOne(try CompositeMetadata.encoded(metadata, using: metadataEncoder))
-    }
-}
-
-
-/// Namespace
-enum Decoders {}
-
-extension Decoders {
-    struct Map<Decoder: DecoderProtocol, Metadata, Data>: DecoderProtocol {
-        var decoder: Decoder
-        var transform: (Decoder.Metadata, Decoder.Data) throws -> (Metadata, Data)
-        mutating func decodedPayload(
-            _ payload: Payload,
-            mimeType: ConnectionMIMEType
-        ) throws -> (Metadata, Data) {
-            let (metadata, data) = try decoder.decodedPayload(payload, mimeType: mimeType)
-            return try transform(metadata, data)
-        }
-    }
-    struct MapMetadata<Decoder: DecoderProtocol, Metadata>: DecoderProtocol {
-        var decoder: Decoder
-        var transform: (Decoder.Metadata) throws -> Metadata
-        mutating func decodedPayload(
-            _ payload: Payload,
-            mimeType: ConnectionMIMEType
-        ) throws -> (Metadata, Decoder.Data) {
-            let (metadata, data) = try decoder.decodedPayload(payload, mimeType: mimeType)
-            return (try transform(metadata), data)
-        }
-    }
-    struct MapData<Decoder: DecoderProtocol, Data>: DecoderProtocol {
-        var decoder: Decoder
-        var transform: (Decoder.Data) throws -> Data
-        mutating func decodedPayload(
-            _ payload: Payload,
-            mimeType: ConnectionMIMEType
-        ) throws -> (Decoder.Metadata, Data) {
-            let (metadata, data) = try decoder.decodedPayload(payload, mimeType: mimeType)
-            return (metadata, try transform(data))
-        }
-    }
-    struct MetadataDecoder<Decoder, MetadataDecoder>: DecoderProtocol where
-    Decoder: DecoderProtocol,
-    Decoder.Metadata: SingleMetadataDecodable,
-    MetadataDecoder: RSocketCore.MetadataDecoder
-    {
-        typealias Metadata = MetadataDecoder.Metadata?
-        typealias Data = Decoder.Data
-        var decoder: Decoder
-        var metadataDecoder: MetadataDecoder
-        mutating func decodedPayload(
-            _ payload: Payload,
-            mimeType: ConnectionMIMEType
-        ) throws -> (Metadata, Data) {
-            let (metadata, data) = try decoder.decodedPayload(payload, mimeType: mimeType)
-            let decodedMetadata = try metadata.decodeMetadata(using: metadataDecoder)
-            return (decodedMetadata, data)
-        }
-    }
-    struct CompositeMetadataDecoder<Decoder, CompositeMetadataDecoder>: DecoderProtocol where
-    Decoder: DecoderProtocol,
-    Decoder.Metadata == [CompositeMetadata],
-    CompositeMetadataDecoder: RSocketCore.CompositeMetadataDecoder
-    {
-        typealias Metadata = CompositeMetadataDecoder.Metadata
-        typealias Data = Decoder.Data
-        var decoder: Decoder
-        var metadataDecoder: CompositeMetadataDecoder
-        mutating func decodedPayload(
-            _ payload: Payload,
-            mimeType: ConnectionMIMEType
-        ) throws -> (Metadata, Data) {
-            let (metadata, data) = try decoder.decodedPayload(payload, mimeType: mimeType)
-            let decodedMetadata = try metadataDecoder.decode(from: metadata)
-            return (decodedMetadata, data)
-        }
-    }
-    struct RootCompositeMetadataDecoder<Decoder>: DecoderProtocol where
-    Decoder: DecoderProtocol,
-    Decoder.Metadata == Foundation.Data?
-    {
-        typealias Metadata = [CompositeMetadata]
-        typealias Data = Decoder.Data
-        var decoder: Decoder
-        var metadataDecoder: RSocketCore.RootCompositeMetadataDecoder
-        mutating func decodedPayload(
-            _ payload: Payload,
-            mimeType: ConnectionMIMEType
-        ) throws -> (Metadata, Data) {
-            let (metadata, data) = try decoder.decodedPayload(payload, mimeType: mimeType)
-            let decodedMetadata = try metadata.map { try metadataDecoder.decode(from: $0) }
-            return (decodedMetadata ?? [], data)
-        }
-    }
-    struct DataDecoder<Decoder, DataDecoder>: DecoderProtocol where
-    Decoder: DecoderProtocol,
-    Decoder.Data == Foundation.Data,
-    DataDecoder: DataDecoderProtocol
-    {
-        typealias Metadata = Decoder.Metadata
-        typealias Data = DataDecoder.Data
-        var decoder: Decoder
-        var dataDecoder: DataDecoder
-        mutating func decodedPayload(
-            _ payload: Payload,
-            mimeType: ConnectionMIMEType
-        ) throws -> (Metadata, Data) {
-            let (metadata, data) = try decoder.decodedPayload(payload, mimeType: mimeType)
-            let decodedData = try dataDecoder.decode(from: data)
-            return (metadata, decodedData)
-        }
-    }
-    struct MultiDataDecoder<Decoder, DataDecoder>: DecoderProtocol where
-    Decoder: DecoderProtocol,
-    Decoder.Metadata == [CompositeMetadata],
-    Decoder.Data == Foundation.Data,
-    DataDecoder: MultiDataDecoderProtocol
-    {
-        typealias Metadata = Decoder.Metadata
-        typealias Data = DataDecoder.Data
-        var decoder: Decoder
-        var dataMIMETypeDecoder: DataMIMETypeDecoder
-        var dataDecoder: DataDecoder
-        var lastSeenDataMIMEType: MIMEType?
-        mutating func decodedPayload(
-            _ payload: Payload,
-            mimeType connectionMIMEType: ConnectionMIMEType
-        ) throws -> (Metadata, Data) {
-            let (metadata, data) = try decoder.decodedPayload(payload, mimeType: connectionMIMEType)
-            if let dataMIMEType = try metadata.decodeFirstIfPresent(using: dataMIMETypeDecoder) {
-                lastSeenDataMIMEType = dataMIMEType
-            }
-            let dataMIMEType = lastSeenDataMIMEType ?? connectionMIMEType.data
-            let decodedData = try dataDecoder.decodeMIMEType(dataMIMEType, from: data)
-            return (metadata, decodedData)
-        }
-    }
-}
-
-extension DecoderProtocol {
-    func map<NewMetadata, NewData>(
-        _ transform: @escaping (Metadata, Data) throws -> (NewMetadata, NewData)
-    ) -> Decoders.Map<Self, NewMetadata, NewData> {
-        .init(decoder: self, transform: transform)
-    }
-    func mapMetadata<NewMetadata>(
-        _ transform: @escaping (Metadata) throws -> NewMetadata
-    ) -> Decoders.MapMetadata<Self, NewMetadata> {
-        .init(decoder: self, transform: transform)
-    }
-    func mapData<NewData>(
-        _ transform: @escaping (Data) throws -> NewData
-    ) -> Decoders.MapData<Self, NewData> {
-        .init(decoder: self, transform: transform)
-    }
-    func decodeMetadata<MetadataDecoder>(
-        using metadataDecoder: MetadataDecoder
-    ) -> Decoders.MetadataDecoder<Self, MetadataDecoder> {
-        .init(decoder: self, metadataDecoder: metadataDecoder)
-    }
-    func decodeMetadata<CompositeMetadataDecoder>(
-        @CompositeMetadataDecoderBuilder metadataDecoder: () -> CompositeMetadataDecoder
-    ) -> Decoders.CompositeMetadataDecoder<Self, CompositeMetadataDecoder> {
-        .init(decoder: self, metadataDecoder: metadataDecoder())
-    }
-    /// unconditionally decodes data with the given `decoder`
-    func decodeData<DataDecoder>(
-        using dataDecoder: DataDecoder
-    ) -> Decoders.DataDecoder<Self, DataDecoder> {
-        .init(decoder: self, dataDecoder: dataDecoder)
-    }
-    /// unconditionally decodes data with the given `decoder`
-    func decodeData<DataDecoder>(
-        dataMIMETypeDecoder: DataMIMETypeDecoder = .init(),
-        @MultiDataDecoderBuilder dataDecoder: () -> DataDecoder
-    ) -> Decoders.MultiDataDecoder<Self, DataDecoder> {
-        .init(decoder: self, dataMIMETypeDecoder: dataMIMETypeDecoder, dataDecoder: dataDecoder())
-    }
-    func useCompositeMetadata(
-        metadataDecoder: RootCompositeMetadataDecoder = .init()
-    ) -> Decoders.RootCompositeMetadataDecoder<Self> where Metadata == Foundation.Data? {
-        .init(decoder: self, metadataDecoder: metadataDecoder)
-    }
-}
-
-struct AnyDecoder<Metadata, Data>: DecoderProtocol {
-    var _decoderBox: _AnyDecoderBase<Metadata, Data>
-    init<Decoder>(
-        _ decoder: Decoder
-    ) where Decoder: DecoderProtocol, Decoder.Metadata == Metadata, Decoder.Data == Data {
-        _decoderBox = _AnyDecoderBox(decoder: decoder)
-    }
-    mutating func decodedPayload(
-        _ payload: Payload,
-        mimeType: ConnectionMIMEType
-    ) throws -> (Metadata, Data) {
-        if !isKnownUniquelyReferenced(&_decoderBox) {
-            _decoderBox = _decoderBox.copy()
-        }
-        return try _decoderBox.decodedPayload(payload, mimeType: mimeType)
-    }
-}
-
-class _AnyDecoderBase<Metadata, Data>: DecoderProtocol {
-    func decodedPayload(
-        _ payload: Payload,
-        mimeType: ConnectionMIMEType
-    ) throws -> (Metadata, Data) {
-        fatalError("\(#function) in \(Self.self) is an abstract method and needs to be overridden")
-    }
-    func copy() -> _AnyDecoderBase<Metadata, Data> {
-        fatalError("\(#function) in \(Self.self) is an abstract method and needs to be overridden")
-    }
-}
-
-final class _AnyDecoderBox<Decoder: DecoderProtocol>: _AnyDecoderBase<Decoder.Metadata, Decoder.Data> {
-    var decoder: Decoder
-    internal init(decoder: Decoder) {
-        self.decoder = decoder
-    }
-    override func decodedPayload(
-        _ payload: Payload,
-        mimeType: ConnectionMIMEType
-    ) throws -> (Decoder.Metadata, Decoder.Data) {
-        try decoder.decodedPayload(payload, mimeType: mimeType)
-    }
-    override func copy() -> _AnyDecoderBase<Decoder.Metadata, Decoder.Data> {
-        _AnyDecoderBox(decoder: decoder)
-    }
-}
-
-extension DecoderProtocol {
-    func eraseToAnyDecoder() -> AnyDecoder<Metadata, Data> {
-        .init(self)
-    }
-}
-
-struct Decoder: DecoderProtocol {
-    init() {}
-    func decodedPayload(_ payload: Payload, mimeType: ConnectionMIMEType) throws -> (Data?, Data) {
-        (payload.metadata, payload.data)
-    }
-}
 
 // MARK: - Payload Encoder
 
@@ -414,7 +69,7 @@ enum Encoders {
     }
     struct MetadataEncoder<Encoder, MetadataEncoder>: EncoderProtocol where
     Encoder: EncoderProtocol,
-    Encoder.Metadata: SingleMetadataEncodable,
+    Encoder.Metadata: MetadataEncodable,
     MetadataEncoder: RSocketCore.MetadataEncoder
     {
         typealias Metadata = MetadataEncoder.Metadata
@@ -437,7 +92,7 @@ enum Encoders {
     
     struct StaticMetadataEncoder<Encoder, MetadataEncoder>: EncoderProtocol where
     Encoder: EncoderProtocol,
-    Encoder.Metadata: SingleMetadataEncodable,
+    Encoder.Metadata: MetadataEncodable,
     MetadataEncoder: RSocketCore.MetadataEncoder
     {
         typealias Metadata = Encoder.Metadata.CombinableMetadata
