+++ Sources/RSocketCore/Extensions/PerStreamDataMIMETypes.swift
@@ -17,19 +17,29 @@
 import Foundation
 import NIO
 
-struct MIMETypeEncoder {
-    private static let defaultWellKnownMimeTypes = Dictionary(
+public struct MIMETypeEncoder {
+    public static let defaultWellKnownMimeTypes = Dictionary(
         uniqueKeysWithValues: MIMEType.wellKnownMIMETypes.lazy.map{ ($0.1, $0.0) }
     )
-    var wellKnownMimeTypes: [MIMEType: WellKnownMIMETypeCode] = defaultWellKnownMimeTypes
+    let wellKnownMimeTypes: [MIMEType: WellKnownMIMETypeCode]
+    public init(
+        wellKnownMimeTypes: [MIMEType : WellKnownMIMETypeCode] = defaultWellKnownMimeTypes
+    ) {
+        self.wellKnownMimeTypes = wellKnownMimeTypes
+    }
     func encode(_ mimeType: MIMEType, into buffer: inout ByteBuffer) throws {
         fatalError("not implemented")
     }
 }
 
-struct MIMETypeDecoder {
-    private static let defaultWellKnownMimeTypes = Dictionary(uniqueKeysWithValues: MIMEType.wellKnownMIMETypes)
-    var wellKnownMimeTypes: [WellKnownMIMETypeCode: MIMEType] = defaultWellKnownMimeTypes
+public struct MIMETypeDecoder {
+    public static let defaultWellKnownMimeTypes = Dictionary(uniqueKeysWithValues: MIMEType.wellKnownMIMETypes)
+    var wellKnownMimeTypes: [WellKnownMIMETypeCode: MIMEType]
+    public init(
+        wellKnownMimeTypes: [WellKnownMIMETypeCode : MIMEType] = defaultWellKnownMimeTypes
+    ) {
+        self.wellKnownMimeTypes = wellKnownMimeTypes
+    }
     func decode(from buffer: inout ByteBuffer) throws -> MIMEType {
         fatalError("not implemented")
     }
@@ -38,56 +48,70 @@ struct MIMETypeDecoder {
 
 // MARK: - Data MIME Type
 
-struct DataMIMETypeEncoder: MetadataEncoder {
-    typealias Metadata = MIMEType
-    var mimeType: MIMEType { .messageXRSocketMimeTypeV0 }
-    var mimeTypeEncoder = MIMETypeEncoder()
-    func encode(_ metadata: Metadata, into buffer: inout ByteBuffer) throws {
+public struct DataMIMETypeEncoder: MetadataEncoder {
+    public typealias Metadata = MIMEType
+    public var mimeType: MIMEType { .messageXRSocketMimeTypeV0 }
+    let mimeTypeEncoder: MIMETypeEncoder
+    public init(mimeTypeEncoder: MIMETypeEncoder = MIMETypeEncoder()) {
+        self.mimeTypeEncoder = mimeTypeEncoder
+    }
+    public func encode(_ metadata: Metadata, into buffer: inout ByteBuffer) throws {
         fatalError("not implemented")
     }
 }
 
-struct DataMIMETypeDecoder: MetadataDecoder {
-    typealias Metadata = MIMEType
-    var mimeType: MIMEType { .messageXRSocketMimeTypeV0 }
-    var mimeTypeDecoder = MIMETypeEncoder()
-    func decode(from buffer: inout ByteBuffer) throws -> Metadata {
-        fatalError("not implemented")
-    }
-}
-
-extension MetadataEncoder where Self == DataMIMETypeEncoder {
+public extension MetadataEncoder where Self == DataMIMETypeEncoder {
     static var dataMIMEType: Self { .init() }
 }
 
-extension MetadataDecoder where Self == DataMIMETypeDecoder {
+public struct DataMIMETypeDecoder: MetadataDecoder {
+    public typealias Metadata = MIMEType
+    public var mimeType: MIMEType { .messageXRSocketMimeTypeV0 }
+    let mimeTypeDecoder: MIMETypeEncoder
+    public init(mimeTypeDecoder: MIMETypeEncoder = MIMETypeEncoder()) {
+        self.mimeTypeDecoder = mimeTypeDecoder
+    }
+    public func decode(from buffer: inout ByteBuffer) throws -> Metadata {
+        fatalError("not implemented")
+    }
+}
+
+public extension MetadataDecoder where Self == DataMIMETypeDecoder {
     static var dataMIMEType: Self { .init() }
 }
 
 // MARK: - Acceptable Data MIME Type
 
-struct AcceptableDataMIMETypeEncoder: MetadataEncoder {
-    typealias Metadata = [MIMEType]
-    var mimeType: MIMEType { .messageXRSocketMimeTypeV0 }
-    var mimeTypeEncoder = MIMETypeEncoder()
-    func encode(_ metadata: Metadata, into buffer: inout ByteBuffer) throws {
+public struct AcceptableDataMIMETypeEncoder: MetadataEncoder {
+    public typealias Metadata = [MIMEType]
+    public var mimeType: MIMEType { .messageXRSocketMimeTypeV0 }
+    let mimeTypeEncoder: MIMETypeEncoder
+    public init(
+        mimeTypeEncoder: MIMETypeEncoder = MIMETypeEncoder()
+    ) {
+        self.mimeTypeEncoder = mimeTypeEncoder
+    }
+    public func encode(_ metadata: Metadata, into buffer: inout ByteBuffer) throws {
         fatalError("not implemented")
     }
 }
 
-struct AcceptableDataMIMETypeDecoder: MetadataDecoder {
-    typealias Metadata = [MIMEType]
-    var mimeType: MIMEType { .messageXRSocketMimeTypeV0 }
-    var mimeTypeDecoder = MIMETypeEncoder()
-    func decode(from buffer: inout ByteBuffer) throws -> Metadata {
-        fatalError("not implemented")
-    }
-}
-
-extension MetadataEncoder where Self == AcceptableDataMIMETypeEncoder {
+public extension MetadataEncoder where Self == AcceptableDataMIMETypeEncoder {
     static var acceptableDataMIMEType: Self { .init() }
 }
 
-extension MetadataDecoder where Self == AcceptableDataMIMETypeDecoder {
+public struct AcceptableDataMIMETypeDecoder: MetadataDecoder {
+    public typealias Metadata = [MIMEType]
+    public var mimeType: MIMEType { .messageXRSocketMimeTypeV0 }
+    let mimeTypeDecoder: MIMETypeEncoder
+    public init(mimeTypeDecoder: MIMETypeEncoder = MIMETypeEncoder()) {
+        self.mimeTypeDecoder = mimeTypeDecoder
+    }
+    public func decode(from buffer: inout ByteBuffer) throws -> Metadata {
+        fatalError("not implemented")
+    }
+}
+
+public extension MetadataDecoder where Self == AcceptableDataMIMETypeDecoder {
     static var acceptableDataMIMEType: Self { .init() }
 }
